backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
pillow==10.4.0
qrcode[pil]==7.4.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
fakeredis==2.20.1
python-dotenv==1.0.0
aiofiles==23.2.1
mypy==1.7.1
types-Pillow==10.1.0.2
types-requests==2.31.0.10
types-redis==4.6.0.11
